tests for Delete User and Users
