+++ routes/admin.js
@@ -4,34 +4,8 @@ const Joi = require('joi');
 const Boom = require('boom');
 const verifyCredentials = require('../models/verifyCredentials');
 const createToken = require('../models/createToken');
-const User = require('../models/user');
+const {deleteUser, returnUsers} = require('../models/admin_helper');
 
-function returnUsers () {
-  return User.find({}).then(users => {
-    if (!users) {
-      const doesNotExistError = new Error('No User Found');
-      throw doesNotExistError;
-    } else {
-      return users;
-    }
-  })
-  .catch(err => {
-    return (Boom.badRequest(err.message));
-  });
-}
-function deleteUser (id) {
-  return User.findOneAndRemove({_id: id}).then(userDeleted => {
-    if (!userDeleted) {
-      const doesNotExistError = new Error('No Such User Found');
-      throw doesNotExistError;
-    } else {
-      return userDeleted;
-    }
-  })
-  .catch(err => {
-    return (Boom.badRequest(err.message));
-  });
-}
 module.exports = [
   {
     method: 'GET',
